+++ CHANGELOG.md
@@ -1,4 +1,21 @@
-## 0.3.9-dev4
+## 0.3.11-dev0
+
+### Enhancements
+
+* **Support Databricks personal access token**
+
+## 0.3.10
+
+### Enhancements
+
+* **Support more concrete FileData content for batch support**
+
+### Fixes
+
+* **Add Neo4J to ingest destination connector registry**
+* **Fix closing SSHClient in sftp connector**
+
+## 0.3.9
 
 ### Enhancements